Git Import Wizard
This is the new Import wizard for importing Projects from Git.
Similar to CVS, it first shows a list of Git Repositories. The user
may select one from the list (the same list as shown in the 
Git Repositories View) and continues from there. Alternatively,
the user can clone or add a Repository to the list.
Then all continues like with the Import Projects... wizard
used in the Git Repositories View.

Bug: 281394

+++ org.eclipse.egit.ui/src/org/eclipse/egit/ui/internal/clone/GitCloneWizard.java
@@ -3,6 +3,7 @@
  * Copyright (C) 2008, Shawn O. Pearce <spearce@spearce.org>
  * Copyright (C) 2008, Marek Zawirski <marek.zawirski@gmail.com>
  * Copyright (C) 2008, Robin Rosenberg <robin.rosenberg@dewire.com>
+ * Copyright (C) 2010, Mathias Kinzler <mathias.kinzler@sap.com>
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -29,73 +30,63 @@ import org.eclipse.egit.ui.internal.repository.RepositoriesView;
 import org.eclipse.jface.dialogs.ErrorDialog;
 import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.jface.operation.IRunnableWithProgress;
-import org.eclipse.jface.viewers.IStructuredSelection;
 import org.eclipse.jface.wizard.Wizard;
 import org.eclipse.jgit.lib.Ref;
 import org.eclipse.jgit.transport.URIish;
 import org.eclipse.osgi.util.NLS;
-import org.eclipse.ui.IImportWizard;
 import org.eclipse.ui.IViewPart;
-import org.eclipse.ui.IWorkbench;
 import org.eclipse.ui.PlatformUI;
 
 /**
  * Import Git Repository Wizard. A front end to a git clone operation.
  */
-public class GitCloneWizard extends Wizard implements IImportWizard {
+public class GitCloneWizard extends Wizard {
 	private RepositorySelectionPage cloneSource;
 
 	private SourceBranchPage validSource;
 
 	private CloneDestinationPage cloneDestination;
 
-	private GitProjectsImportPage importProject;
+	String alreadyClonedInto;
 
-	public void init(IWorkbench arg0, IStructuredSelection arg1) {
+	/**
+	 * The default constructor
+	 */
+	public GitCloneWizard() {
 		setWindowTitle(UIText.GitCloneWizard_title);
 		setDefaultPageImageDescriptor(UIIcons.WIZBAN_IMPORT_REPO);
 		setNeedsProgressMonitor(true);
 		cloneSource = new RepositorySelectionPage(true, null);
 		validSource = new SourceBranchPage(cloneSource);
 		cloneDestination = new CloneDestinationPage(cloneSource, validSource);
-		importProject = new GitProjectsImportPage() {
-			@Override
-			public void setVisible(boolean visible) {
-				if (visible) {
-					if (cloneDestination.alreadyClonedInto == null) {
-						if (performClone(false))
-							cloneDestination.alreadyClonedInto = cloneDestination
-									.getDestinationFile().getAbsolutePath();
-					}
-					setProjectsList(cloneDestination.alreadyClonedInto);
-				}
-				super.setVisible(visible);
-			}
-		};
 	}
 
 	@Override
 	public boolean performCancel() {
-		if (cloneDestination.alreadyClonedInto != null) {
-			if (MessageDialog
-					.openQuestion(getShell(), UIText.GitCloneWizard_abortingCloneTitle,
+		if (alreadyClonedInto != null) {
+			File test = new File(alreadyClonedInto);
+			if (test.exists()
+					&& MessageDialog.openQuestion(getShell(),
+							UIText.GitCloneWizard_abortingCloneTitle,
 							UIText.GitCloneWizard_abortingCloneMsg)) {
-				deleteRecursively(new File(cloneDestination.alreadyClonedInto));
+				deleteRecursively(new File(alreadyClonedInto));
 			}
 		}
 		return true;
 	}
 
 	private void deleteRecursively(File f) {
-		for (File i : f.listFiles()) {
-			if (i.isDirectory()) {
-				deleteRecursively(i);
-			} else {
-				if (!i.delete()) {
-					i.deleteOnExit();
+		File[] children = f.listFiles();
+		if (children != null)
+			for (File i : children) {
+				if (i.isDirectory()) {
+					deleteRecursively(i);
+				} else {
+					if (!i.delete()) {
+						i.deleteOnExit();
+					}
 				}
 			}
-		}
 		if (!f.delete())
 			f.deleteOnExit();
 	}
@@ -105,21 +96,16 @@ public class GitCloneWizard extends Wizard implements IImportWizard {
 		addPage(cloneSource);
 		addPage(validSource);
 		addPage(cloneDestination);
-		addPage(importProject);
 	}
 
 	@Override
 	public boolean canFinish() {
-		return cloneDestination.isPageComplete()
-				&& !cloneDestination.showImportWizard.getSelection()
-				|| importProject.isPageComplete();
+		return cloneDestination.isPageComplete();
 	}
 
 	@Override
 	public boolean performFinish() {
-		if (!cloneDestination.showImportWizard.getSelection())
-			return performClone(true);
-		return importProject.createProjects();
+		return performClone(false);
 	}
 
 	boolean performClone(boolean background) {
@@ -139,6 +125,7 @@ public class GitCloneWizard extends Wizard implements IImportWizard {
 		final String remoteName = cloneDestination.getRemote();
 
 		workdir.mkdirs();
+
 		if (!workdir.isDirectory()) {
 			final String errorMessage = NLS.bind(
 					UIText.GitCloneWizard_errorCannotCreate, workdir.getPath());
@@ -160,7 +147,9 @@ public class GitCloneWizard extends Wizard implements IImportWizard {
 
 		final CloneOperation op = new CloneOperation(uri, allSelected,
 				selectedBranches, workdir, branch, remoteName);
-		importProject.setGitDir(op.getGitDir());
+
+		alreadyClonedInto = workdir.getPath();
+
 		if (background) {
 			final Job job = new Job(NLS.bind(UIText.GitCloneWizard_jobName, uri
 					.toString())) {
@@ -196,8 +185,11 @@ public class GitCloneWizard extends Wizard implements IImportWizard {
 							throws InvocationTargetException,
 							InterruptedException {
 						op.run(monitor);
+						if (monitor.isCanceled())
+							throw new InterruptedException();
 					}
 				});
+
 				RepositorySelectionPage.saveUriInPrefs(uri.toString());
 				RepositoriesView.addDir(op.getGitDir());
 				if (view != null)
